feat: add custom component theme switcher
