+++ src/pages/Layout/Header.tsx
@@ -3,11 +3,7 @@ import logo from '../../assets/color-logo.svg';
 import { CustomButton } from 'components/CustomButton';
 import { useDispatch, useSelector } from 'react-redux';
 import type { AppDispatch } from '../../redux/store';
-import {
-  selectIsLoggedIn,
-  selectUser,
-  selectUserTheme,
-} from '../../redux/users/usersSlice';
+import { selectIsLoggedIn, selectUser } from '../../redux/users/usersSlice';
 import CabinetSVG from '../../assets/CabinetSVG';
 import { openDialog } from '../../redux/dialogs/dialogSlice';
 import {
@@ -21,8 +17,7 @@ import { FaHeart, FaRegHeart } from 'react-icons/fa';
 import { RxCross2, RxHamburgerMenu } from 'react-icons/rx';
 import { cn } from 'components/lib/utils';
 import { useState } from 'react';
-import { changeThemeThunk } from 'src/redux/users/usersOperations';
-import { updateUserLocally } from '../../redux/users/usersSlice';
+import ThemeSwitch from 'components/ThemeSwitch';
 
 export const Header = () => {
   const [isMobileMenuOpen, setMobileMenuOpen] = useState(false);
@@ -32,20 +27,6 @@ export const Header = () => {
   const user = useSelector(selectUser);
   const favoriteAnimals = useSelector(selectFavoriteAnimals);
   const favoritesCount = favoriteAnimals.length;
-
-  const currentTheme = useSelector(selectUserTheme);
-  const newTheme: 'light' | 'dark' = currentTheme === 'dark' ? 'light' : 'dark';
-  const handleToggleTheme = () => {
-    if (isLoggedIn) {
-      dispatch(changeThemeThunk({ theme: newTheme }));
-    } else {
-      dispatch(updateUserLocally({ theme: newTheme }));
-      document.documentElement.classList.remove(
-        currentTheme as 'light' | 'dark'
-      );
-      document.documentElement.classList.add(newTheme);
-    }
-  };
 
   return (
     <>
@@ -109,13 +90,7 @@ export const Header = () => {
             </li>
           </ul>
           <div className="flex items-center gap-28">
-            <CustomButton
-              styleType="whiteButton"
-              onClick={handleToggleTheme}
-              className="rounded-full w-100"
-            >
-              Theme: {currentTheme === 'dark' ? '🌙' : '☀️'}
-            </CustomButton>
+            <ThemeSwitch />
             <NavLink to="/favorite" className="relative" end>
               {favoritesCount > 0 ? (
                 <FaHeart size={32} className="text-error-input" />
@@ -128,7 +103,6 @@ export const Header = () => {
                 </span>
               )}
             </NavLink>
-
             {isLoggedIn ? (
               <div className="flex flex-col items-center max-h-[54px]">
                 <Avatar
